Working code for listing locations

+++ RentalCarsAPI/RentalCarsAPI/Models/Domain/RentalLocation.cs
@@ -14,8 +14,8 @@ namespace RentalCarsAPI.Models.Domain
         public string State { get; set; }
         public string Zip { get; set; }
         public string Phone { get; set; }
-        public float Lat { get; set; }
-        public float Long { get; set; }
+        public double Lat { get; set; }
+        public double Long { get; set; }
         public DateTime DateCreated { get; set; }
         public DateTime DateModified { get; set; }
     }
